sql_chatbot/requirements.txt
fastapi
uvicorn
openai
python-dotenv
mysql-connector-python
jinja2
